[refactor] 예매 진입 상태 비교 로직 공용화 #363
